Setting Headers, timers
