+++ src/components/create-meeting/create-meeting-header/create-meeting-header.component.tsx
@@ -1,9 +1,11 @@
 import React, { FC } from "react";
 import { HelpCircle, Video } from "react-feather";
-import { getCurrentTime, getDateTime } from "../../../utils/time.utills";
+import { getDateTime } from "../../../utils/time.utills";
 import { Signin } from "./signin.component";
 import logo from "../../../assets/imgs/logo.svg";
+import { useTime } from "../../../hooks/use-time.hook";
 export const CreateMeetingHeader: FC = () => {
+  const currentTime = useTime();
   return (
     <header className="w-10/12 text-highlight flex justify-between items-center mx-auto">
       <section className="py-5">
@@ -11,7 +13,7 @@ export const CreateMeetingHeader: FC = () => {
       </section>
       <section className="text-sm my-2 items-center flex">
         <p className="sm:mx-5 mx-0">
-          {getCurrentTime()} &bull; {getDateTime()}
+          {currentTime} &bull; {getDateTime()}
         </p>
         {/* <HelpCircle className="text-sm cursor-pointer mx-3 text-primary" />
         <Video className="text-sm cursor-pointer mx-3 text-primary" /> */}
